Added Template Files for Non Nutil Inputs
These files are templates that can be used for the non Nutil inputs. It is recommended to create a copy of the file at the level you need and then add columns with your data. If you need different regions, feel free to add to or make your own file. Just make sure your region column has valid ABA abbreviations.